requirements.txt
pandas==1.2.3
plotly==4.14.3
seaborn==0.11.1
matplotlib==3.4.1
streamlit==0.76.0
joblib==0.16.0
numpy==1.20.1
Pillow==8.2.0
scikit-learn
